refactor: Action 하나만 사용하도록 수정
