--- src/main/java/chess/controller/room/RoomAction.java
@@ -1,11 +1,0 @@
-package chess.controller.room;
-
-import java.util.List;
-
-@FunctionalInterface
-public interface RoomAction {
-    RoomAction EMPTY = ignore -> {
-    };
-
-    void execute(final List<String> commands);
-}
